ingestion/requirements.txt
sentence-transformers
faiss-cpu
python-dotenv
pydantic-settings
PyGithub